# contracts/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "AddAlertNoteRequest": {
        "properties": {
          "note": {
            "minLength": 1,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "note"
        ],
        "title": "AddAlertNoteRequest",
        "type": "object"
      },
      "AlertDetailResponse": {
        "properties": {
          "history": {
            "items": {
              "$ref": "#/components/schemas/AlertHistoryResponse"
            },
            "title": "History",
            "type": "array"
          },
          "record": {
            "$ref": "#/components/schemas/AlertRecordResponse"
          }
        },
        "required": [
          "history",
          "record"
        ],
        "title": "AlertDetailResponse",
        "type": "object"
      },
...
      "AlertRecordResponse": {
        "properties": {
          "acknowledged_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "acknowledged_by_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged By User Id"
          },
          "alert_id": {
            "title": "Alert Id",
            "type": "string"
          },
          "current_value": {
            "title": "Current Value",
            "type": "number"
          },
          "engine": {
            "$ref": "#/components/schemas/DatabaseEngine"
          },
          "instance_id": {
            "title": "Instance Id",
            "type": "string"
          },
          "last_evaluated_at": {
            "title": "Last Evaluated At",
            "type": "string"
          },
          "metric_name": {
            "title": "Metric Name",
            "type": "string"
          },
          "opened_at": {
            "title": "Opened At",
            "type": "string"
          },
          "owner_assigned_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Assigned At"
          },
          "owner_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner User Id"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "rule_name": {
            "title": "Rule Name",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          }
        },
        "required": [
          "alert_id",
          "acknowledged_at",
          "acknowledged_by_user_id",
          "current_value",
          "engine",
          "instance_id",
          "last_evaluated_at",
          "metric_name",
          "opened_at",
          "owner_assigned_at",
          "owner_user_id",
          "resolved_at",
          "rule_id",
          "rule_name",
          "severity",
          "status",
          "threshold"
        ],
        "title": "AlertRecordResponse",
        "type": "object"
      },
      "AlertRuleResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "engine": {
            "$ref": "#/components/schemas/DatabaseEngine"
          },
          "instance_ids": {
            "items": {
              "type": "string"
            },
            "title": "Instance Ids",
            "type": "array"
          },
          "metric_name": {
            "title": "Metric Name",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "operator": {
            "title": "Operator",
            "type": "string"
          },
          "overrides": {
            "items": {
              "$ref": "#/components/schemas/RuleOverrideResponse"
            },
            "title": "Overrides",
            "type": "array"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          }
        },
        "required": [
          "created_at",
          "enabled",
          "engine",
          "instance_ids",
          "metric_name",
          "name",
          "operator",
          "overrides",
          "rule_id",
          "severity",
          "threshold"
        ],
        "title": "AlertRuleResponse",
        "type": "object"
      },
      "AlertStatus": {
        "enum": [
          "acknowledged",
          "open",
          "resolved"
        ],
        "title": "AlertStatus",
        "type": "string"
      },
      "AssignAlertOwnerRequest": {
        "properties": {
          "owner_user_id": {
            "title": "Owner User Id",
            "type": "string"
          }
        },
        "required": [
          "owner_user_id"
        ],
        "title": "AssignAlertOwnerRequest",
        "type": "object"
      },
...
      "CreateRuleRequest": {
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "engine": {
            "$ref": "#/components/schemas/DatabaseEngine"
          },
          "instance_ids": {
            "items": {
              "type": "string"
            },
            "title": "Instance Ids",
            "type": "array"
          },
          "metric_name": {
            "title": "Metric Name",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "operator": {
            "$ref": "#/components/schemas/RuleOperator"
          },
          "overrides": {
            "items": {
              "$ref": "#/components/schemas/RuleOverrideRequest"
            },
            "title": "Overrides",
            "type": "array"
          },
          "severity": {
            "$ref": "#/components/schemas/RuleSeverity"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          }
        },
        "required": [
          "engine",
          "metric_name",
          "name",
          "operator",
          "severity",
          "threshold"
        ],
        "title": "CreateRuleRequest",
        "type": "object"
      },
      "DatabaseEngine": {
        "enum": [
          "mysql",
          "oracle"
        ],
        "title": "DatabaseEngine",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RuleOperator": {
        "enum": [
          "gt",
          "gte",
          "lt",
          "lte"
        ],
        "title": "RuleOperator",
        "type": "string"
      },
...
      "RuleSeverity": {
        "enum": [
          "critical",
          "warning"
        ],
        "title": "RuleSeverity",
        "type": "string"
      },
...
      "UpdateRuleRequest": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "engine": {
            "$ref": "#/components/schemas/DatabaseEngine"
          },
          "instance_ids": {
            "items": {
              "type": "string"
            },
            "title": "Instance Ids",
            "type": "array"
          },
          "metric_name": {
            "title": "Metric Name",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "operator": {
            "$ref": "#/components/schemas/RuleOperator"
          },
          "overrides": {
            "items": {
              "$ref": "#/components/schemas/RuleOverrideRequest"
            },
            "title": "Overrides",
            "type": "array"
          },
          "severity": {
            "$ref": "#/components/schemas/RuleSeverity"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          }
        },
        "required": [
          "enabled",
          "engine",
          "metric_name",
          "name",
          "operator",
          "severity",
          "threshold"
        ],
        "title": "UpdateRuleRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "db-monitor-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/alerts": {
      "get": {
        "operationId": "list_alerts_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AlertStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RuleSeverity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "instance",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Instance"
            }
          },
          {
            "in": "query",
            "name": "opened_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opened After"
            }
          },
          {
            "in": "query",
            "name": "opened_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opened Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertRecordResponse"
                  },
                  "title": "Response List Alerts Alerts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Alerts"
      }
    },
...
    "/alerts/rules": {
      "get": {
        "operationId": "list_rules_alerts_rules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertRuleResponse"
                  },
                  "title": "Response List Rules Alerts Rules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Rules"
      },
      "post": {
        "operationId": "create_rule_alerts_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rule"
      }
    },
    "/alerts/rules/{rule_id}": {
      "get": {
        "operationId": "get_rule_alerts_rules__rule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rule"
      },
      "put": {
        "operationId": "update_rule_alerts_rules__rule_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Rule"
      }
    },
    "/alerts/{alert_id}": {
      "get": {
        "operationId": "get_alert_alerts__alert_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Alert"
      }
    },
    "/alerts/{alert_id}/acknowledge": {
      "post": {
        "operationId": "acknowledge_alert_alerts__alert_id__acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acknowledge Alert"
      }
    },
    "/alerts/{alert_id}/notes": {
      "post": {
        "operationId": "add_alert_note_alerts__alert_id__notes_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAlertNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Alert Note"
      }
    },
    "/alerts/{alert_id}/owner": {
      "put": {
        "operationId": "assign_alert_owner_alerts__alert_id__owner_put",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignAlertOwnerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Alert Owner"
      }
    },
...
  }
}